+++ Civi/DataProcessor/Factory.php
@@ -153,6 +153,8 @@ class Factory {
     $this->addFilter('contact_in_group_filter', new Definition('Civi\DataProcessor\FilterHandler\ContactInGroupFilter'), E::ts('Contact in Group filter'));
     $this->addFilter('contact_with_tag_filter', new Definition('Civi\DataProcessor\FilterHandler\ContactWithTagFilter'), E::ts('Contact has Tag filter'));
     $this->addFilter('contact_has_membership', new Definition('Civi\DataProcessor\FilterHandler\ContactHasMembershipFilter'), E::ts('Contact has Membership filter'));
+    $this->addFilter('contact_has_event_registration_of_type', new Definition('Civi\DataProcessor\FilterHandler\ContactHasEventRegistrationOfType'), E::ts('Contact has event registration of type'));
+    $this->addFilter('contact_has_contribution_with_status', new Definition('Civi\DataProcessor\FilterHandler\ContactHasContributionWithStatus'), E::ts('Contact has contribution with status'));
     $this->addFilter('worldregion_filter', new Definition('Civi\DataProcessor\FilterHandler\WorldRegionFilter'), E::ts('World Region Filter'));
     $this->addFilter('contact_type_filter', new Definition('Civi\DataProcessor\FilterHandler\ContactTypeFilter'), E::ts('Contact Type filter'));
     $this->addFilter('permission_to_view_contact', new Definition('Civi\DataProcessor\FilterHandler\PermissionToViewContactFilter'), E::ts('Permission to view contact'));